Split off `ParticipantService` from `StudyService`.
