+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/StudiesCodeSamples.kt
@@ -2,7 +2,6 @@ package dk.cachet.carp.studies
 
 import dk.cachet.carp.common.EmailAddress
 import dk.cachet.carp.common.UUID
-import dk.cachet.carp.deployment.application.DeploymentService
 import dk.cachet.carp.deployment.application.DeploymentServiceHost
 import dk.cachet.carp.deployment.domain.StudyDeploymentStatus
 import dk.cachet.carp.deployment.infrastructure.InMemoryAccountService
@@ -13,6 +12,8 @@ import dk.cachet.carp.protocols.domain.StudyProtocolSnapshot
 import dk.cachet.carp.protocols.domain.devices.AnyMasterDeviceDescriptor
 import dk.cachet.carp.protocols.domain.devices.Smartphone
 import dk.cachet.carp.protocols.domain.tasks.ConcurrentTask
+import dk.cachet.carp.studies.application.ParticipantService
+import dk.cachet.carp.studies.application.ParticipantServiceHost
 import dk.cachet.carp.studies.application.StudyService
 import dk.cachet.carp.studies.application.StudyServiceHost
 import dk.cachet.carp.studies.domain.ParticipantGroupStatus
@@ -20,6 +21,7 @@ import dk.cachet.carp.studies.domain.StudyStatus
 import dk.cachet.carp.studies.domain.users.AssignParticipantDevices
 import dk.cachet.carp.studies.domain.users.Participant
 import dk.cachet.carp.studies.domain.users.StudyOwner
+import dk.cachet.carp.studies.infrastructure.InMemoryParticipantRepository
 import dk.cachet.carp.studies.infrastructure.InMemoryStudyRepository
 import dk.cachet.carp.test.runBlockingTest
 import kotlin.test.*
@@ -29,7 +31,7 @@ class StudiesCodeSamples
 {
     @Test
     fun readme() = runBlockingTest {
-        val studyService: StudyService = createStudiesEndpoint()
+        val (studyService, participantService) = createEndpoints()
 
         // Create a new study.
         val studyOwner = StudyOwner()
@@ -44,7 +46,7 @@ class StudiesCodeSamples
 
         // Add a participant.
         val email = EmailAddress( "participant@email.com" )
-        val participant: Participant = studyService.addParticipant( studyId, email )
+        val participant: Participant = participantService.addParticipant( studyId, email )
 
         // Once all necessary study options have been configured, the study can go live.
         if ( studyStatus is StudyStatus.Configuring && studyStatus.canGoLive )
@@ -59,17 +61,22 @@ class StudiesCodeSamples
             val participation = AssignParticipantDevices( participant.id, setOf( patientPhone.roleName ) )
             val participantGroup = setOf( participation )
 
-            val groupStatus: ParticipantGroupStatus = studyService.deployParticipantGroup( studyId, participantGroup )
+            val groupStatus: ParticipantGroupStatus = participantService.deployParticipantGroup( studyId, participantGroup )
             val isInvited = groupStatus.studyDeploymentStatus is StudyDeploymentStatus.Invited // True.
         }
     }
 
 
-    private fun createStudiesEndpoint(): StudyService =
-        StudyServiceHost( InMemoryStudyRepository(), createDeploymentService() )
+    private fun createEndpoints(): Pair<StudyService, ParticipantService>
+    {
+        val studyRepo = InMemoryStudyRepository()
+        val studyService = StudyServiceHost( studyRepo )
 
-    private fun createDeploymentService(): DeploymentService =
-        DeploymentServiceHost( InMemoryDeploymentRepository(), InMemoryAccountService() )
+        val deploymentService = DeploymentServiceHost( InMemoryDeploymentRepository(), InMemoryAccountService() )
+        val participantService = ParticipantServiceHost( studyRepo, InMemoryParticipantRepository(), deploymentService )
+
+        return Pair( studyService, participantService )
+    }
 
     /**
      * This is the protocol created in ProtocolsCodeSamples.readme().
