+++ carp.studies.core/src/commonTest/kotlin/dk/cachet/carp/studies/application/StudyServiceHostTest.kt
@@ -1,10 +1,6 @@
 package dk.cachet.carp.studies.application
 
-import dk.cachet.carp.deployment.application.DeploymentServiceHost
-import dk.cachet.carp.deployment.infrastructure.InMemoryAccountService
-import dk.cachet.carp.deployment.infrastructure.InMemoryDeploymentRepository
 import dk.cachet.carp.studies.infrastructure.InMemoryStudyRepository
-import dk.cachet.carp.studies.domain.StudyRepository
 
 
 /**
@@ -12,13 +8,5 @@ import dk.cachet.carp.studies.domain.StudyRepository
  */
 class StudyServiceHostTest : StudyServiceTest
 {
-    override fun createService(): Pair<StudyService, StudyRepository>
-    {
-        val repo = InMemoryStudyRepository()
-        val accountService = InMemoryAccountService()
-        val deploymentService = DeploymentServiceHost( InMemoryDeploymentRepository(), accountService )
-        val service = StudyServiceHost( repo, deploymentService )
-
-        return Pair( service, repo )
-    }
+    override fun createService(): StudyService = StudyServiceHost( InMemoryStudyRepository() )
 }
